refactor(cli): extract parsePositiveInt to cli/utils/option-parsers
